refactor(framework): Move AppIoServicer
